First commit, not sure what this repo is about yet. Collection of my work so far.
In the `Merged` folder we have the converted dataset from MCScript and
COIN (1st task), using `xml2json`. We also have a script to clean and
merge everything into a single file (we'll split and shuffle things
later).

In `test.py` we have a simple test that outputs some data in format
we'll use for the training of the model: (passage, question, answer).