Reorganises data files, dataset on a csv, implements baseline.
Decided to use csv instead of json, as that's easier to parse in the
model code. I'm currently (in the baseline) using the `data_joined.csv`
file and just splitting by the '#'. Might change it in the future and
just split by '|', doesn't really matter.

Baseline is a LSTM for each of the passage, question, answer, then
concatenated into a feed-forward with 2 outputs. I was trying to make
it work with just one output, but haven't succeeded yet.

The model seems to be bad, of course. I didn't let it train for too
long, because the current machine is weak.

All in all, AllenNLP seems like a lot of work but everything makes sense
when it's done. Now (well, after I change the things mentioned above) I'm
going to delete the comments (from the original tutorial) and write some
of my own.