Got CUDA to work, removed the comments on that section.
The CUDA problems had nothing to do with the code. It was a driver issue
on my side.

For future reference, this worked: `sudo modprobe -vv nvidia`.
Apparently I have to run this after the driver is updated, sometimes.

It's working now, but of course training is still very slow. I'll try
reducing the LSTM sizes (300 hidden_dim is too much). And I'm running
out of memory, apparently.

AMMEND: Actually, 300 input is necessary because of the embeddings
(which are 300d). I'm going to download a smaller GloVe and see how it
goes.

I actually thought I was using 300 hidden because the FFN is also using
300, but that's because it's 3*100 (LSTM is using 100). I'll experiment
with different sizes and see if I can at least train locally without
running out of memory.