Move files to specific folders. Splits the script into multiple modules.
The data folder was renamed. The scripts were kept in there (I don't
want to put them together with the neural net files).

The old `baseline.py` was split into various files, as per last commit
message. I think they are fine now, but I'm going to refactor the main
script to be more readable and organised.

Re: .gitignore: added `__pycache__` to it.