Extract methods from `baseline` into a separate utilities file.
The `util` module now has several functions that were extracted from the
main script.

The main script now builds the layers and the model, than passes that
to a more generic `train_model` function (that does the right thing for
our problem, but in a model-agnostic way). It's also responsible for
outputting some predictions and testing the loading code.

The `train_model` function is still responsible for saving the model,
although I'm considering if that's appropriate. My current answer is
yes, because otherwise I'd have to return both the vocabulary and the
model from the function, but that's not a big problem.

Or would I? Maybe I can get the vocabulary directly from the model.
Anyway, I'll explore this. I'll try to get this architecture right
before moving on.