Registers Model, Predictor and Reader. Also renames them.
I decided to register everything, even though I'm not using the config
approach as of yet. Whilst doing that, I realised that the reader and
the predictor are dataset specific, not a general QA. So I renamed them.

About the Reader I'm not sure. I named it McScriptReader, but it's not even
appropriate to the original format (which is some XML mess), but rather
to my own processed input. I'm keeping the name, anyway.

As for the Predictor, it's task rather than dataset specific, much like
the Model is. But I named the Predictor McScript but not the Model. I
might come back to this, but I'm not spending too long on naming.
This'll probably change in the future.

Next steps: build new model, based on Baseline but with attention.
I'll also experiment with the Tokenizer class from AllenNLP, see if I
can get NER and POS tags from it.

When I do (either from it or some external source, I'll check Yuanfudao),
I'll build yet another classifier using them alongside the embeddings in
the input (again, following Yuanfudao).

After that, all that will remain to reproduce Yuanfudao is the
ConceptNet embeddings, I think. And some tweaks to architecture as well
(dimensions, dropout, etc).