Registers Model, Predictor and Reader. Also renames them.
I decided to register everything, even though I'm not using the config
approach as of yet. Whilst doing that, I realised that the reader and
the predictor are dataset specific, not a general QA. So I renamed them.

About the Reader I'm not sure. I named it McScriptReader, but it's not even
appropriate to the original format (which is some XML mess), but rather
to my own processed input. I'm keeping the name, anyway.

As for the Predictor, it's task rather than dataset specific, much like
the Model is. But I named the Predictor McScript but not the Model. I
might come back to this, but I'm not spending too long on naming.
This'll probably change in the future.

Next steps: build new model, based on Baseline but with attention.
I'll also experiment with the Tokenizer class from AllenNLP, see if I
can get NER and POS tags from it.

When I do (either from it or some external source, I'll check Yuanfudao),
I'll build yet another classifier using them alongside the embeddings in
the input (again, following Yuanfudao).

After that, all that will remain to reproduce Yuanfudao is the
ConceptNet embeddings, I think. And some tweaks to architecture as well
(dimensions, dropout, etc).

NOTE: I'll try to keep using the generic names for things
(TextFieldEmebdder, Seq2VecEncoder, Seq2SeqEncoder, whatever is the
generic for Attention when I get there). This way, after I've finished
reproducing Yuanfudao, I can change little code to replace GloVe with
ELMo and BERT, and maybe even experiment with a Transformer instead of
LSTMs (or even a GRU), and other types of attention.

Eventually I might try the Jsonnet configuration approach anyway, but it
makes me a bit uneasy for some reason.

Oh, just realised that I can streamline my current Embeddings and LSTM
instantiation. And implement some configuration of my own, probably
using JSON itself because it's easy.