Extracted the GloVe and LSTM instantiations.
Extracted the code into separate functions and moved them to `util`. I
might eventually create another module `layers` and keep them there.

There's more work to be done, here. Ideally I would be able to have a
method that a reads a config (with paths, dimensions, types of layers)
and instantiate a model from that, and then train it.

I realise that's very similar to what the AllenNLP config does, but I
still don't want to use that.