Now splitting dataset to a test set and evaluating on it.
I'm also fixing a bug that was taking me to use the wrong index
to the logit tensor. I thought that label 0 would be index 0 (and
vice-versa), but I was wrong apparently. I'm still using string labels
because that seemed more natural, but I'm not entire sure how it works
with integer labels (they are supposed to be 0-indexed, so it should be
a straightfoward mapping). Regardless, I'm keeping it this way because
there's no actual difference anyway.

In the end there's actually no way to use a binary label if I want to
use AllenNLP's Accuracy. They only accept CategoricalAccuracy, which
forces me into a categorical classification task even if I only have two
labels. PyTorch would let me do it just fine (they have
BinaryCrossEntropy), but AllenNLP doesn't.

I'm considering if I should open an issue on the AllenNLP repository.
Maybe I'm missing something. Or maybe I'm not, and they'll tell me to
keep using this and just do what I'm doing (get the probability of the
positive class).