Implements proper binary classification. Preliminary config options.
So, I finally cracked the binary classification thing (instead of
categorical with 2 classes).

I was fighting AllenNLP, because PyTorch has the appropriate loss function
for the problem, but AllenNLP didn't have an accuracy function for
binary, just for categorical. So I implemented the metric myself, which
was pretty straightforward (I legit copy-pasted the code from
Categorical and changed the correct items computation, and some
dimensionality checks).

Now I can finally use the final layer with one output.

I'm also now printing the whole instance in the examples evaluation. And
I instead of changing the paths and dimensions, I have both with an if
controlling them. I'm slowly converging to a proper configuration
scheme, and I'm not yet sure if I'll use AllenNLP's. I'll look into it,
though.