Add config for `medium`. Changes path for pre-processed input.
I had it on the `data` folder originally, but then realised it's going
to be HUGE and moved to to `External`.

I need to figure out a better way to config this external folder. Later
(it's a TODO).

Oh, re: huge. I had a MemoryError locally trying to save the
pre-processed `data` file, so I did it on Colab.