Adds RNN configuration and CLI config for data to use.
Now we can use GRUs as well (instead of just LSTM), and we can also use
BiLSTMs (and BiGRUs), and stack them.

These are now offered as options on top of the script too. The config
option is now passable via CLI, so we don't have to commit different
versions for the Colab notebook.