Implements AttentiveClassifier, using attention, and some configuration.
The attention model is similar to the one from Yuanfudao (it's pretty
much identical except for the last layer, where it uses another bilinear
and sums the logits). We have Bilinear for Passage-Question and Linear
for self-question and self-answer.

The bilinear is straightforward, matching with what the library expects,
but the self-attention was a bit weird. I have to use a `combination`
parameter that ignores the second vector (as, well, it isn't a vector).

But then it works, and it does what the one from Yuanfudao does anyway
(which is only multiplying the weight vector by the input).

I also added yet another bit of configuration (in the script itself, no
arguments just yet) for the model. There's also the possibility to use
GRU instead of the LSTM, but I don't even know if it's practical.

Oh, and the model currently takes the same RNN parameters for the three
text layers. Yuanfudao uses different parameters for each, and they're
dependent on a lot of things. I might have issues keeping it
parametrised in the same way. I'll try to keep it as-is, but when I
actually implement Yuanfudao I'll need to drop the idea.

Now I'm going to reproduce the attention before the LSTM, and then
replace my Linear with the Bilinear from Yuanfudao. And then it's
feature engineering time.