Report about Yuanfudao's output layer and how to use dropout
This commit doesn't have any new code in it, it's just an entry on the
research journal (which is how I'm treating these commits anyway). I had
them on their respective tasks on Todoist, but I felt it would be better
to have it here.

The output isn't Attention, it's two Linear layers.  For passage *
answer, we pass the passage to the Linear, and we get a vector of size
answer back. Then we multiply this vector with the answer, then sum.
The same is done for question * answer.

These two sums are summed together and fed through sigmoid.

This does make some sense because it creates an interaction between the
answer and the other two, whilst my concatenation just lumps them
together.

I'll try that tomorrow.

Re: dropout. Yeah, it is easy, just use `torch.nn.Dropout` directly.
Initialise it with the probability we want, then just call it on the
layer.

Should be easy. I can also refer to Yuanfudao.