Add proper customisation of optimiser on `train_model`
I was accepting the object itself, but since it needs the model to be
instantiated, it was kind of useless. Now it accepts a function that
takes a `Model` and returns an `Optimizer`.

So you can build a lambda (or a local function, as autopep8 seems to
prefer) and configurate the optimiser as you wish. I'm testing with
Adamax with the default parameters (the learning rate of 2e-3 sounds a
bit weird, but it's also the default in Yuanfudao).

Training was unstable, might've been (suprised Pikachu) because of the
high learning rate (lr=0.1). I might try SGD again to see if that was
it.

Tried with the small dataset and it seemed to learn fine, but it's too
small to be reliable. I'm training locally on the medium now, let's see
how that goes. I'll put the large one to train on Colab later.