Batch size and number of epochs per config now.
Since I realised that I'm changing it wrt why I'm running (small for
debugging, medium for some idea of whether the model is performing
better, large for actually training the damn thing), it doesn't make
much sense to have a global configuration for each.

As to model performance, I'm still running on Colab. But my training on
local (with batch_size = 128, which might be too big for 5000 instances;
I'll update on a future commit if size = 32 is better) shows some size
of overfitting already. I was going to work on dropout anyway, this
looks like an interesting reminder.