First Attentive Reader implementation.
Now I'm passing the encoder for each field separately. This allows me to
have a stacked RNN for the passage and shallow ones for the rest.

Also, the modification here was just on the output: instead of
concatenating the three outputs and feeding into a FNN, I have a
bilinear multiplication of the Passage-Answer weighted states and the
Answer weighted states. This gives me a vector, which I sum and feed
into a sigmoid.

The article does a few things differently, that I could try but would
require more changes to the text. They would be quite simple, actually,
I might do them anyway.

These are:
    - I'm using Seq2Seq for Question and Answer, and then weighting the
    hidden states according to self-attention. The paper model uses
    Seq2Vec for these, and only uses a Seq2Seq for the text
    - They use GRUs instead of LSTMs
    - Apparently every layer is shallow, so no stacking for the text
    either

To change this I had to create a whole new model class, which is why I
haven't yet.

Preliminary tests (with _medium_) indicates that this model of combining
output is actually good. Of course, I'd only know after training on the
large model anyway.

By the way, result from training the previous version: got 60% accuracy
on testing after 200 epochs with the large dataset. It's decent, maybe
this output implementation can give a better result.

I'm going to implement the "proper" version and see how they compete.