Implement proper Attentive Reader. CLI argument for the model to use.
I decided to implement as described on the paper (somewhat). Here's what
it does:

1. Encodes passage using GRU into a sequence of hidden states (Seq2Seq)
2. Encodes questin and answer into a vector using GRU (Seq2Vec)
3. Computes a weight vector using passage-question bilinear attention
4. Calculates passage vector using a weighted sum with vector from 3
5. Another bilinear between the weighted passage and the answer
6. This returns a scalar (vector * vector) logit
7. Sigmoid of that scalar is the output probability

Now I'm also accepting the model as a CLI argument, as I want to run
both models simultaneously in Colab.

I'm going to experiment whether my model (with Yuanfudao-style weighting
of the hidden states instead of just using the output) is better than
the AttentiveReader with the large dataset on Colab.