Report on Yuanfudao implementation (output layer).
So, the work in Attentive Readershows this is pretty straightfoward. I
just need another linear layer and the multiplication. It's really just
duplicating some of the code we have now.

I'm currently running the code on Colab to see if my current model is
better than my implementation of the AttentiveReader. If mine is better
I'll just leave the AttentiveReader as-is. If not, I'll keep adding the
modifications to both.