Implementer accuracy based on questions answered correctly.
Before I was calculating the accuracy based on the model predicting
whether a wrong answer was wrong and a right answer was right. This is
still the metric (and loss) that I use during training, but on the
evaluation of the test set, now I calculate based on the question.

This means that the model doesn't need to know that a wrong answer is
wrong. All it needs is to be more confident that the right answer is
actually right.

Whilst this works for evaluation. I'm wondering about the consequences
that this has for training. I feel like I'm training them for a subtly
different task to what I'm evaluating them on. Perhapse the model as
described in the baseline should work better.

That would involve many changes, though. Off the top of my head:

- Change data to have the two answers as part of the same tuple. While
I'm at it, I'd keep the ID's as well, as they're important for the
submission later on.
- Change the DatasetReader accordingly.
- Change the McScriptPredictor accordingly.
- Change the problem from a Binary to Categorical. This would undo all
the work I've done, but now I actually have a reason to use Categorical.
It doesn't change much theoretically, but now my outputs make sense in a
Categorical perspective.
- However, this would imply in some more elaborate output behaviour. I'd
have to basically compute two outputs in the same way I'm doing now,
then put them together and softmax them. Which works, but is a bit ugly.
Anyway, I _have_ to use the same layers for both answers, nothing else
makes sense.

I'm doing all this in a separate branch, as it's quite a bit of work. If
all goes well (and performance improves, which is a big one), I'll merge
back to master.