Re-implements models in terms of both answers. Re-implements input reading.
The models now accept two answers and return a 2-vector of the
probabilities of each being the correct one. Now I'm using
CategorialAccuracy and CrossEntropyLoss again, and I deleted the
BinaryAccuracy class I had before.

I also decided to stick to the JSONs as input, as converted from the
xml2json program. This means that I handle everything in the reading
class, no cleaning scripts will be run before (just xml2json to convert
them, and even that could be revisited in the future. I just hate XML).

I might have to write another merging script to combine the datasets
anyway, to get a big JSON file for training to get my results. But that
should be straightfoward, as I'd just replace the IDs, the rest would
remain the same. I'm already seeing it: ready both JSONs into dicts,
update the IDs and write them again.

This means that I'm also deleting the CSV files and related scripts.

All this accomplished what I wanted: performance seems to be better now.
I still have to run them on Colab, though, and see.

There are two things I'm not happy with:
    - Whatever the instance has is passed to the `forward` function of
    the model, and the IDs (Passage and Question) are useless.
    - I'm currently using `stack` to combine the two probabilities and
    then softmax. It feels weird and patchy, but I'm not sure there's a
    better way of doing it.