Apply dropout to embeddings and RNN layers.
For the embeddings we have a Dropout object, as the functional interface
needs an explicit `training` parameter. I consulted some AllenNLP models
and they used the Dropout class, so I did the same.

For the RNN layers I'm using the `dropout` parameter, that applies the
dropout to each layer except for the last. So I'm also using another
Dropout object (to allow for different parameters, although I'm using
the same right now) after the output.

As for the values, the AttentiveReader is using the value described in
the paper (0.5), and AttentiveClassifier is using the same value as
Yuanfudao (0.4).