Changes tokens to lowercase. Adds some configuration options.
I added the option to fine-tune the embeddings, but that seemed to cause
overfitting so I'm disabling it (option still there, though).

The lowercasing of the tokens doesn't seem to make that big of a
difference. We'll see with the large instances.

I'm still curious about the fine-tuning of the embeddings, I thought the
dropout would've been enough to contain the overfitting.