Implements Yuanfudao's attention in the input layer.
This was called `Sequence Attention`, where we have two sequences u and
v, and we weigh v by scores from uWv.

I'm doing it slightly different from the class in Yuanfudao, as they do
the weighted sum in their class, but since I'm inheriting from
AllenNLP's attention, I'm just outputting the scores themselves (and the
base class deals with the mask normalisation for me), and calculating
the sum in Model::forward. This leads to a bit of duplicated code, but
is more consistent with the rest of the code.