More options for the experiments.
I'm testing some approaches to fight the overfitting:

- Changing LSTM hidden layer size (96 to 64)
- Whether or not to fine tune
- GloVe dimensionality (300d or 100d)
- Which optimiser to use
    * The idea here is that they might interact differently with the
    fine tuning

What we have now:

- The model does achieve higher accuracy with the fine-tuning, but stops
learning really fast.
- In fact, training loss and validation loss diverge
- Without fine-tuning, they trend towards the same path, but very slowly

So I'm running these experiments to see what happens. We'll have 50
epochs on the whole training dataset. Besides that, we have 2*2*4*2 = 32
different scenarios to test for. Each takes around 45 minutes.

This is going to take a while. I'm going to work on the NER and POS
embeddings whilst that's going on.