Enables POS and NER as input to the model.
For now, I'm only computing POS and NER for the passage and POS for the
question (following Yuanfudao). Adding NER for the question and both for
the answer is easy (and I might attempt that, but I imagine Yuanfudao
tried), but it does seem a bit overkill.

Next step is creating two more embedding layers, one for each, and use
them alongside the text embeddings in the model.