Finalises NER and POS embeddings as part of Attentive.
I had a small bug where there was a confusion on namespaces. I thought
the namespace would be the string I gave it, but PosTagIndexer and
NerTagIndexer have their defaults (which have a `_tokens` suffix that
mine didn't).

Now everything is using the default, so no problems (the `forward`
method complains, as the keys in the tensor dicts are different).

I also faced a smill hiccup in that I changed the namespaces in the
code, but as I was loading a pre-processed file, they weren't doing
anything (training worked but evaluation, which alwasy used the new code,
failed).

Now everything works. I'm still testing the dimensions and optimisation
algorithms in Colab (which should take a long time still). After I get
results from that, I'm running this to see if there's any visible
improvement.