Removes configuration options (for now).
I needed those options to test different combinations of parameters, but
now that it's done¹, I'm removing it because it was a hack. I'll
eventually work on proper configuration (likely not AllenNLP's, as I
have too much custom stuff).

¹: As for the result of test, it didn't run completely and I couldn't
get the TensorBoard files because the instance was stopped before I got
there. I could read the output, though, and they showed the same
pattern.

The solution here is indeed to fine-tune the embeddings, as it leads to
an increase to the overall accuracy. However, learning stops very
quickly. Without fine-tuning it does too, but with lower accuracy.

I'm considering what can I do to prevent this. Higher dropout,
recurent_dropout (which would involve changing the LSTM class), trying
to use Yuanfudao's top-10 fine-tuning (which I have no clue how to).

One of my attempts was to reduce the dimension of the embeddings and the
LSTM, but that apparently doesn't make that much difference.