Implements BERT as the encoding layer for the model.
I had to bang my head a little bit because of lack of documentation and
examples, but I made it work.

Caveats: BERT uses WordPieces, so I had to change the Tokeniser and the
indexer. This is not particularly configurable right now (the reader
receives them as parameters, but the training code doesn't have a switch
for this), so it involves changing the code directly. The model itself
didn't need changing.

The model is also huge. The output dimension of BERT is 768, which
causes some of the intermediate layers to be huge (the answer encoder,
for example, is 3 * dim). This makes the model horrendously slow. I
don't think there's any way of reducing this (besides putting it through
an FNN first).

Because of this, I've only ran it in the small dataset. Not even medium
is feasible locally. I actually have no clue how this performs.

Next steps:
- Run this is Colab and see what happens
- Use BERT in Baseline and Attentive Reader too
- ConceptNet Embeddings