Implements Relation Embeddings in Attentive. Some cleanup and documentation.
Cleanup according to some warnings from Pylint. I couldn't fix some (it
complains about number of parameters and variables, and some stupid
"different arguments from overriden method" stuff).

Also wrote some small docstrings for all functions. They're not very
useful, being that small, but better than nothing in most cases. I'm not
paying much attention to documentation right now, but I should do that
before I drop this project (before exams).

There are two things I'm not satisfied for now:
1. Configurability: especially after introducing BERT, everything is too
coupled. I believe I'd have implement a new Reader to be able to
decouple from BERT, give all the indexing and tokenising going on
2. Speaking of tokenising, I can't use the BERT-specific tokeniser
because it doesn't give me the NER and POS tag. On the other hand, using
Spacy is not ideal, since BERT has a tokeniser of its own. But if I
split and use different tokenisers for each (BertTokenizer for the text
itself, Spacy for the rest), then they do different things and the
sequences have different lengths, which of course doesn't work.

I most likely won't be able to address 2 without major changes (like
tokenising with Bert and then manually parsing NER and POS, but that's a
lot harder than this sounds [it's probably sequence-aware, so I'd have
to do some voodo to convert between the two]).

For 1 though, I might be able to parametrise the Reader to avoid having
to write two separate variations. And that's not even the beginning of
the story: now I have many parameters scattered accross the code. Some
are explicit as global variables in the `run` script, maybe some others
aren't. I'd like to have a way to configure them all, with the defaults
as they are now.

Whether I'll use CLI arguments or a configuration file, I'm not sure
yet. I probably won't address this until I'm back full-time on this
(read: after exams), since the current setup is a bit of work but does
what I need it to.

The next steps are:
- Implement a scheduler to reduce the LR on plateus (should be
straightforward, I already have the class for this (ReduceLROnPlateau,
descriptive name right?), I should be able to instantiate this and pass
to the traning object. This will probably be the next commit.
- Investigating the remainig features for Yuanfudao, and whether they're
worth the hassle.
- Dirty hack to be able to use GloVe again for some experiments
- Run experiments on Attentive (which is now almost Yuanfudao) and
Attentive+BERT