Move embedding functions to layers. Configuration for Glove/Bert.
I had the functions that returned configured embeddings (glove, basic
and bert) in the utils module for some reason. Now they're on the layers
module. I wasn't sure because they're not classes, but they're no
utility functions either.

I found a hacky way to configure between Glove and Bert too. The reader
can just receive the indexer (SingleIdIndexer for Glove, BertIndexer for
Bert), I didn't need to rewrite it (that concern came from when I was
using the Bert WordSplitter, which ended up being a bad idea anyway).

The configuration really is on the build_attentive function (that
decides on which embedding to send to the model) and on the train_model
function (which decides which token_indexer to pass to the reader). I'm
still not very comfortable doing the input reading on the training
function directly, so I might change that. I would take a lot of logic
somewhere else, though.