Allows proper configuration between BERT and GloVe. Solves a few bugs.
The main bug was: for some reason while training the Vocabulary creates
a namespace "bert" that is then persisted by `Vocabulary::save_to_files`.
However, the very own `Vocabulary::from_files` can't load the persisted
vocabulary, because that namespace lacks an OOV token, but at the same
time isn't on the non-padded namespaces list.

The solution is to pickle/un-pickle it instead. It does the right
thing, but could be version-specific. I might open an Issue on GitHub
for this in the future.

As to "proper configuration", I extracted the data loading bit from the
`train_model` function and now I'm passing the dataset itself to it. I
also had to pass the reader itself to the testing functions instead of
initialising it there, to save me the hassle of replicating the logic to
creating it.