Fixes tests. Replaces path joining by proper os.path.join.
Tests now instantiate their own readers, which should spot save/load-related
bugs in the future. They still receive the layers as parameters,
however. I'm starting to wonder if it wouldn't be better to just pickle
the whole model instead of loading the weights.

A con of this is that I'd be dependant on the AllenNLP version, breaking
changes would break the pickle but potentially not the weights. I'll
leave it as-is for now.

As for the paths: ideally I'd use Path (pathlib) instead of this. This
seemed like a quicker change, as I'm not familiar with pathlib.