Enables parsing in the tokeniser. Some output for testing.
I noticed that the program looks like it hangs during testing, as
there's no indication what it's doing. I added some output to explain.

Enabled the parsing in the tokeniser. Now we have the Token lemma, which
we can then use to compute the handcrafted features from Yuanfudao.

Also introduced `wikiwords` as a dependency. This will be used to
calcualte the term frequency of a word, again for Yuanfudao.