Implements Yuanfudao's Handcrafted features (all of them).
There are two differences:
- The calculation for the TF is different, as the original doesn't make
much sense (it's scaled by 0.1, but there's a +10 inside the log). We're
just doing log(1+f_w) instead of 0.1*log(10 + f_w).
- For the co-occurence features, we have to do it for both answers,
because we're outputting the softmax over the two. The alternative would
be to go through the entire network twice, one for each, and softmax at
the end. I'm not doing that.