Attempt to make BERT work.
BERT has a maximum size on the number of wordpieces a single string can
have. The AllenNLP implementation truncates if the string exceeds that
number.

However, the POS and NER weren't truncating the same way. I truncated
before passing the string to the indexers so I wouldn't have the problem
of mismatching sequence sizes.

Of course, the best way would be to do no truncation whatsoever, but
that would involve a lot more work. I'm going to attempt this now, and
work more on BERT later.