Reimplements Linear Self Attention.
I thought that the AllenNLP implementation was the problem, but there is
something causing the bias term to go into NaN. This isn't happening
with my re-implementation of the Bilinear Attention, but is with the
reimplementation of the Self Attention.

I decided to remove the bias term, and the errors stopped. I must,
however, test the performance on the full dataset, and see if anything
changes. If not, I'm going to leave this as-is, as I have more things to
do than chase NaN bugs.