Implements a Simple BERT model, changes parameters for the Transformer.
This Simple Bert is just BERT followed by a pooler, which merely gets
the first token embedding (the CLS token). This is then followed by a
linear network into the probability outputs.

It must be noted that this involved a different processing of the input:
we now have '[CLS] <Question> [SEP] <Passage> [SEP] <Answer0> [SEP]
<Answer1>'. Instead of using BERT for each part, we combine and use a
single BERT.

I decided to just add this on top of what the other models need, which
does increase the memory consumption, but wasn't that much.

AllenNLP has a slightly more complex Pooler, that I might test in the future.

Also adjusted paramters for the Transformer, as the basic parameters
from the paper were causing a Memory Overflow.