Fix bug in BERT reader (using question twice in input string).
I also had manually inserted the [CLS] token, but it appears that the
indexer does that already. This doesn't seem to make the SimpleBert
model any better, though.

The next modification for SimpleBert will be using the chunks method
from Run_Squad. The problem with this is that we have to find a way to
combine the answers, as we effectively split a big passage into
multiple, each coupled with the entire query + answers.

I am still unsure as to how to do this, so I'm probably stick to the old
idea for now (separately encode the passage, query and answers), and
explore the Transformer, and maybe other encodings (Relationship
Embeddings, Numberbatch).