Separates the Readers for each type of model.
This eliminates the useless features for some of the models, such as
BERT strings for the models that don't use that format directly. This
will also hopefully reduce the memory footprint.

I still had to keep the "useless" paramters for the SimpleBert reader,
as those are used in sorting the batches. I could change that, but it's
the only model that needs this, so I'd rather just keep everything
there.