Implements AdvancedBert and HierarchicalBert.
AdvancedBert is now SimpleBert with an encoder layer on top of the
Pooler, instead of the simple max from SimpleBert.

HierarchicalBert is a a hierarchical Seq2Vec on top of BERT. First we
encode the sentence representations with a Seq2Vec model, which yields a
sequence of vectors. This sequence is then encoded into a document
representation by another Seq2Vec. This is thne sent into a Dense layer.