Fixes sliding windows
I was calculating the window sizes based on the character length. For
some reason this was causing some windows to be larger than the maximum
of wordpieces, so now I'm doing it based on the tokens directly.

I'm still not positive this is going to work. I might just reduce the
maximum size altogether.