Implements AdvancedBert and HierarchicalBert with Attention.
On model of HAN (Hierarchical Attention Networks). The HierarchicalBert
is basically a full reproduction of HAN with BERT as embeddings (I don't
remember what the original uses). AdvancedBert just has attention after
the only encoder it has.

This involved using the standard LinearSelfAttention (such as in TriAN).
The encoders (the single in AdvancedBert, the sentence and document in
HierarchicalBert) were previously Seq2Vec. They were converted to
Seq2Seq, the states are weighted with self-attention, and summed. Let's
see how it performs.