Reduces number of epochs. Print save path at the end.
The save path is to avoid scrolling all the way back up to find what was
the model being trained. The number of epochs is to reduce the total
training time, as it seems that the models converge (validation-wise)
around that number of epochs, or at least show the learning pattern.