Adds an option to use CNN as the combination of the relation embeddings.
Before I was using the mean of the vectors (as the sequence shouldn't
matter). Now I also have the option of using a CnnEncoder.

Also fixed a bug where I wasn't using seq_to_seq to map the Encoder over
the Relation sentences. It works for BertPooler (as it's just a Dense
layer), but doesn't work for RNNs.