Changes the way reflexivity works in ConceptNet.
Before I just had the database contain both sides of the relation, but
as we're encoding them as sentences, the order makes a diference now.
So instead we have both checks on the loop.

Also removed the case when no relation is found, since that's likely not
going to happen, considering how many places we get relations from. If I
ever manage to get a different head for each relation source, then it
might have to come back.

Also removed the check for same word, as I believe it doesn't make that
much sense, and might make sense in some cases.