Re-implements Transformer, implements RelationalTransformer.
The AllenNLP "transformer" was different from the paper one, so I did it
again. I also implemented my own type, which combines two inputs (src
and kb).

I implemented a RelationTransformerModel, that uses the Transformer to
encode the sentences (both for text and kb), and then joins them using
the RelationalTransformer.