Implements error analsysis and XLNet.
The XLNet implementation is a bit rough, there are probably a few bugs
to be uncovered when I run this on the cluster. But it works on the
Small, I'm just not sure if it's 100%. It doesn't look to be learning,
but it's a small dataset anyway.

As for eror analysis, I'm just separating the evaluation dataset in
parts (text/commonsense) to see how each model behaves.