Changes how global arguments are handled.
common.py has a global ARGS variable that all its function use.
What I _should_ do is pass `args` as a parameter and get rid of
the global altogether... But that's too much work.

Instead, I'll keep the global, but have a module set-function,
`common.set_args`, that sets the value. The end results is
exactly this same, but this looks cleaner.

Overall I'm happy with the final state of the argument processing,
especially considering how it all began. This is still not ideal,
but it's alright.