Adds `graph.py`. It generates a Graph of the models.
Graph is manually generated based on the perceived hierarchy.
This is not directly related to the code (althoug the classes
are defined here after all), but I found it easier to do this
in code rather than draw it somewhere else. As such, I'm
including it with the rest of the code.